app/requirements.txt
asgiref==3.6.0
Django==4.1.6
django-cleanup==6.0.0
django-cors-headers==3.14.0
django-crispy-forms==1.14.0
django-filter==22.1
django-multiselectfield==0.1.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pdfkit==1.0.0
Pillow==9.4.0
psycopg2==2.9.4
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
